service geupdatet

+++ controllers/ServiceController.js
@@ -27,7 +27,7 @@ async function addService(req, res) {
 
 async function updateService(req, res) {
     try {
-        const { id } = req.params;
+        const { id } = req.query; // Lesen der ID aus den Abfrageparametern
         const updates = req.body;
         const [updatedRowsCount] = await Service.update(updates, { where: { id } });
         if (updatedRowsCount === 0) {
@@ -40,9 +40,15 @@ async function updateService(req, res) {
     }
 }
 
+
+
+
 async function deleteService(req, res) {
     try {
-        const { id } = req.params;
+        const { id } = req.query; // Lesen der ID aus den Abfrageparametern
+        if (!id) {
+            return res.status(400).json({ error: 'Die ID wurde nicht bereitgestellt!' });
+        }
         const deletedRowsCount = await Service.destroy({ where: { id } });
         if (deletedRowsCount === 0) {
             return res.status(404).json({ error: 'Dienst nicht gefunden!' });
@@ -53,9 +59,10 @@ async function deleteService(req, res) {
         return res.status(500).json({ error: 'Ein Fehler ist beim Löschen des Dienstes aufgetreten!' });
     }
 }
+
 async function getServiceById(req, res) {
     try {
-        const { id } = req.params;
+        const { id } = req.query; // Änderung: Lesen der ID aus den Abfrageparametern
         const service = await Service.findByPk(id);
         if (!service) {
             return res.status(404).json({ error: 'Dienst nicht gefunden!' });
@@ -66,6 +73,9 @@ async function getServiceById(req, res) {
         return res.status(500).json({ error: 'Ein Fehler ist beim Abrufen des Dienstes aufgetreten!' });
     }
 }
+
+
+
 module.exports = {
     getAllServices,
     addService,
